+++ CHANGELOG.md
@@ -11,6 +11,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Spacing in habits search header
 - Habit completion card typography & whitespace
 
+### Fixed:
+- Keyboard brightness when following system dark/light modes
+
 ## [0.8.369] - 2023-06-06
 ### Changed:
 - Improve habit completion chart colors